Another one

+++ src/modules/TraceRouteModule.cpp
@@ -624,7 +624,7 @@ void TraceRouteModule::drawFrame(OLEDDisplay *display, OLEDDisplayUiState *state
                         int lastGoodBreak = -1;
                         bool lineComplete = false;
 
-                        for (int i = 0; i < remaining.length(); i++) {
+                        for (int i = 0; i < static_cast<int>(remaining.length()); i++) {
                             char ch = remaining.charAt(i);
                             String testLine = tempLine + ch;
